fix restricted scan behavior
Previously, scanning with an API key could cause you to prematurely
believe the scan was finished. If you got a page for which all results
got filtered out (due to restricted permissions), you would
get back an empty page, which is also the signal for "end-of-scan".

Now, the server will keep grabbing pages until it either hits the end
of the scan or gets enough results post-filter to satisfy your limit.